:recycle: Refactor todo list outside home view
