+++ lib/ui/todo_list/todo_list_view.dart
@@ -1,26 +1,26 @@
 import "package:flutter/material.dart";
+import "package:gitdone/ui/_widgets/dropdown_filter_chip.dart";
 import "package:gitdone/ui/_widgets/todo_card.dart";
-import "package:gitdone/ui/home/home_viev_view_model.dart";
-import "package:gitdone/ui/home/widgets/dropdown_filter_chip.dart";
+import "package:gitdone/ui/todo_list/todo_list_view_model.dart";
 import "package:provider/provider.dart";
 
 /// The home view of the app, displayed after login.
-class HomeView extends StatefulWidget {
-  /// Creates a new instance of [HomeView].
-  const HomeView({super.key});
+class TodoListView extends StatefulWidget {
+  /// Creates a new instance of [TodoListView].
+  const TodoListView({super.key});
 
   @override
-  State<HomeView> createState() => _HomeViewState();
+  State<TodoListView> createState() => _TodoListViewState();
 }
 
-class _HomeViewState extends State<HomeView> {
+class _TodoListViewState extends State<TodoListView> {
   @override
   Widget build(final BuildContext context) => Scaffold(
     body: Padding(
       padding: const EdgeInsets.only(top: 16),
       child: ChangeNotifierProvider(
-        create: (_) => HomeViewViewModel()..loadTodos(),
-        child: Consumer<HomeViewViewModel>(
+        create: (_) => TodoListViewModel()..loadTodos(),
+        child: Consumer<TodoListViewModel>(
           builder: (final context, final model, _) => Column(
             children: [
               _buildSearchField(model),
@@ -38,7 +38,7 @@ class _HomeViewState extends State<HomeView> {
     floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
   );
 
-  Widget _buildSearchField(final HomeViewViewModel model) => Padding(
+  Widget _buildSearchField(final TodoListViewModel model) => Padding(
     padding: const EdgeInsets.symmetric(horizontal: 16),
     child: TextField(
       decoration: const InputDecoration(
@@ -50,7 +50,7 @@ class _HomeViewState extends State<HomeView> {
     ),
   );
 
-  Widget _buildFilterRow(final HomeViewViewModel model) => Padding(
+  Widget _buildFilterRow(final TodoListViewModel model) => Padding(
     padding: const EdgeInsets.only(right: 16),
     child: Row(
       mainAxisAlignment: MainAxisAlignment.end,
@@ -91,7 +91,7 @@ class _HomeViewState extends State<HomeView> {
     onUpdate: (final item) => onUpdate(item.value),
   );
 
-  Widget _buildTodoList(final HomeViewViewModel model) {
+  Widget _buildTodoList(final TodoListViewModel model) {
     if (model.isEmpty) {
       return const Expanded(
         child: Center(child: Text("No Issues found in this repository")),
